Fix focus issue in dialog with focustrap
When treeselectjs is used inside a dialog that has a focus trap, the
event is a focus event and the target is the dialog element instead of
the element that was focused for real.
This lead to isInsideClick being false even when the click is done
inside of the modal / dialog, thus closing the list.